refactor(series): move series-card component and add search title
- Move SeriesCardComponent to a shared level within search
- Add title to search route
- Add commit-msg skill for reference
