+++ src/app/series/search/components/series-card/series-card.component.html
